backend/requirements.txt
# FastAPI Core
fastapi==0.128.1
uvicorn[standard]==0.40.0
python-dotenv==1.2.1
httpx==0.28.1
pydantic==2.12.5
slowapi==0.1.9
python-multipart==0.0.9

# AI
google-generativeai

# Vector DB (cloud-based, zero disk storage)
pinecone-client==5.0.0

# LangChain (minimal)
langchain-google-genai

# TTS (free, no API limits)
gTTS

# Core
sentence-transformers
